// ui/src/client/workflowImport/types.gen.ts
export type CallDispositionCodes = {
    /**
     * Disposition Codes
     */
    disposition_codes?: Array<string>;
};
// ...
export type MakeImportAndCreateResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Status
     */
    status: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Workflow Definition
     */
    workflow_definition: {
        [key: string]: unknown;
    };
    /**
     * Current Definition Id
     */
    current_definition_id: number | null;
    /**
     * Template Context Variables
     */
    template_context_variables?: {
        [key: string]: unknown;
    } | null;
    call_disposition_codes?: CallDispositionCodes | null;
    /**
     * Total Runs
     */
    total_runs?: number | null;
    /**
     * Workflow Configurations
     */
    workflow_configurations?: {
        [key: string]: unknown;
    } | null;
    /**
     * Version Number
     */
    version_number?: number | null;
    /**
     * Version Status
     */
    version_status?: string | null;
    /**
     * Warnings
     */
    warnings?: Array<string>;
};
// ...
export type N8nImportAndCreateResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Status
     */
    status: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Workflow Definition
     */
    workflow_definition: {
        [key: string]: unknown;
    };
    /**
     * Current Definition Id
     */
    current_definition_id: number | null;
    /**
     * Template Context Variables
     */
    template_context_variables?: {
        [key: string]: unknown;
    } | null;
    call_disposition_codes?: CallDispositionCodes | null;
    /**
     * Total Runs
     */
    total_runs?: number | null;
    /**
     * Workflow Configurations
     */
    workflow_configurations?: {
        [key: string]: unknown;
    } | null;
    /**
     * Version Number
     */
    version_number?: number | null;
    /**
     * Version Status
     */
    version_status?: string | null;
    /**
     * Warnings
     */
    warnings?: Array<string>;
};
// ...
export type SkillImportAndCreateResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Status
     */
    status: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Workflow Definition
     */
    workflow_definition: {
        [key: string]: unknown;
    };
    /**
     * Current Definition Id
     */
    current_definition_id: number | null;
    /**
     * Template Context Variables
     */
    template_context_variables?: {
        [key: string]: unknown;
    } | null;
    call_disposition_codes?: CallDispositionCodes | null;
    /**
     * Total Runs
     */
    total_runs?: number | null;
    /**
     * Workflow Configurations
     */
    workflow_configurations?: {
        [key: string]: unknown;
    } | null;
    /**
     * Version Number
     */
    version_number?: number | null;
    /**
     * Version Status
     */
    version_status?: string | null;
    /**
     * Warnings
     */
    warnings?: Array<string>;
    /**
     * Suggested Template Variables
     */
    suggested_template_variables?: Array<string>;
};
// ...
export type ZapierImportAndCreateResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Status
     */
    status: string;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Workflow Definition
     */
    workflow_definition: {
        [key: string]: unknown;
    };
    /**
     * Current Definition Id
     */
    current_definition_id: number | null;
    /**
     * Template Context Variables
     */
    template_context_variables?: {
        [key: string]: unknown;
    } | null;
    call_disposition_codes?: CallDispositionCodes | null;
    /**
     * Total Runs
     */
    total_runs?: number | null;
    /**
     * Workflow Configurations
     */
    workflow_configurations?: {
        [key: string]: unknown;
    } | null;
    /**
     * Version Number
     */
    version_number?: number | null;
    /**
     * Version Status
     */
    version_status?: string | null;
    /**
     * Warnings
     */
    warnings?: Array<string>;
};